Move the inspector into components subfolder
